+++ Binary-Search/easy/69-sqrtx/sqrtx.py
