+++ Two-Pointers/medium/1621-number-of-subsequences-that-satisfy-the-given-sum-condition/number-of-subsequences-that-satisfy-the-given-sum-condition.py
@@ -1,0 +1,16 @@
+class Solution:
+    def numSubseq(self, nums: List[int], target: int) -> int:
+        nums.sort()
+
+        left = 0
+        right = len(nums) - 1
+        res = 0
+        MOD = 10**9 + 7
+        while left <= right:
+            if nums[left] + nums[right] <= target:
+                res += pow(2, right - left, MOD)
+                res = res % MOD
+                left += 1
+            else:
+                right -= 1
+        return res
